Simplify GitRepo end-to-end tests (#3682)
* Run gitrepo end-to-end tests without loop

This uses 3 separate test bodies, making it easier to keep track of
which test cases are executed and which are not.

* Remove check on error message from logs

That check does not belong here, and may result in flaky tests.

* Move repoName const to suite level

This clarifies that it is a constant used by multiple test drivers.
It takes the opportunity to rename occurrences of `repoName` to
`localRepoName` in `e2e/single-cluster/gitrepo_test.go`, to make the
override of that constant more explicit.