Monitor known_hosts entries through updatecli (#3680)
* Add updatecli pipeline updating known-hosts config map

This will allow that config map to be synchronised as often as updatecli
pipelines run, hence daily, with updated `known_hosts` entries from
listed providers.

* Run script as external file

This prevents syntax errors seen when inlining the script.

* Replace `title` with `name`

`title` as a field name is deprecated in favour of `name`, as stated by
updatecli output.