Simplify collision detection in `fleet apply`
Relying on the fact that `fleet apply` does not support creating HelmOps
bundles at this point, there is no need to generalise collision
detection to any mismatch in HelmOps state between a possibly existing
bundle and the one to be created or updated.

Instead, checking whether a HelmOps bundle already exists is enough to
establish a collision.