Detect bundle naming collision (#3718)
* Detect bundle naming collision in HelmOps reconciler

When building a HelmOps bundle, the reconciler now aborts with an error
if a non-HelmOps bundle with the same name exists in the same namespace.

* Detect bundle naming collision in `fleet apply`

When creating or updating a bundle, `fleet apply` checks the spec of any
existing bundle with the same name in the same namespace. If one bundle
is a HelmOps bundle while the other is not, `fleet apply` will abort the
update and output an error message.

* Simplify collision detection in `fleet apply`

Relying on the fact that `fleet apply` does not support creating HelmOps
bundles at this point, there is no need to generalise collision
detection to any mismatch in HelmOps state between a possibly existing
bundle and the one to be created or updated.

Instead, checking whether a HelmOps bundle already exists is enough to
establish a collision.