Poll Helm from scheduled job
With scheduling logic in place, Fleet is now able to poll the Helm
registry pointed to by a HelmOp resource, and to update the bundle
created from that HelmOp resource accordingly.

This still requires additional status updates, such as a last-polled
timestamp on the HelmOp resource.

Polling failures are tested through integration tests, while the happy
case is validated in end-to-end tests, which may change in a future
commit. In the meantime, templating of test Kubernetes resources now
relies on `text/template` instead of `html/template`, to prevent version
constraints such as `< 1.0` from being escaped as `&lt; 1.0`, which
Fleet would then interpret as invalid semver constraints.