Update polling time upon successful polling
When a polling job is able to successfully retrieve a new Helm chart
version and apply it to the bundle, the job updates the parent HelmOp's
status with a timestamp indicating when Helm was last polled.

This could be improved to run that update as long as polling Helm
succeeds, even if subsequent steps, such as updating the bundle itself,
fail.