Skip HelmOp version update from reconciler if polling applies
Ensuring that a HelmOp's version fields are only updated from a single
place limits risks of race conditions. The polling job will be
responsible for updating those fields if polling applies to the HelmOp.

Status updates run by the HelmOps reconciler, on the other hand, are now
done using a `Patch` operation, more lightweight than a full `Update`,
and skipped if the patch is empty.