Skip deployment of HelmOps bundle with version constraint
When a bundle is of HelmOps type and still bears a version constraint,
for instance because the initial HelmOps polling job has not yet updated
the bundle with its chart version resolved from Helm, the bundle
reconciler will not create bundle deployments. Instead, it will return
an error and feed it into the bundle status' `Ready` condition, for
better visibility in the Rancher UI.

This prevents bundle deployments from being created with a version which
does not resolve to a single Helm version, which may lead to
inconsistencies in deployments of that bundle among target clusters.