Poll Helm repository from HelmOps reconciler (#3784)
* Schedule polling jobs for HelmOps

The HelmOps controller is now able to schedule a polling job for a
HelmOp resource meeting both of the following conditions:
- a version constraint resolving to a non-static version
- a non-zero polling interval

A job will poll the Helm registry pointed to by a HelmOp resource, and
update the bundle created from that HelmOp resource accordingly.

Polling failures are tested through integration tests, while the happy
case is validated in end-to-end tests.
In the meantime, templating of test Kubernetes resources now
relies on `text/template` instead of `html/template`, to prevent version
constraints such as `< 1.0` from being escaped as `&lt; 1.0`, which
Fleet would then interpret as invalid semver constraints.

* Generate event from Helm polling job

The Helm polling job now generates events when it successfully obtains a
new version for a Helm chart, and when it fails for various reasons as
well.

* Update polling time

When a polling job is able to successfully retrieve a new Helm chart
version and apply it to the bundle, the job updates the parent HelmOp's
status with a timestamp indicating when Helm was last polled.

The polling job also updates the latest polling timestamp in the HelmOp
status when operations which run after polling the Helm registry, such as
getting a bundle, fail. This should make troubleshooting easier.

* Skip HelmOp version update from reconciler if polling applies

Ensuring that a HelmOp's version fields are only updated from a single
place limits risks of race conditions. The polling job will be
responsible for updating those fields if polling applies to the HelmOp.

Status updates run by the HelmOps reconciler, on the other hand, are now
done using a `Patch` operation, more lightweight than a full `Update`,
and skipped if the patch is empty.

* Skip deployment of HelmOps bundle with version constraint

When a bundle is of HelmOps type and still bears a version constraint,
for instance because the initial HelmOps polling job has not yet updated
the bundle with its chart version resolved from Helm, the bundle
reconciler will not create bundle deployments. Instead, it will return
an error and feed it into the bundle status' `Ready` condition, for
better visibility in the Rancher UI.

This prevents bundle deployments from being created with a version which
does not resolve to a single Helm version, which may lead to
inconsistencies in deployments of that bundle among target clusters.