Set HelmOps Accepted condition when no error happened (#3822)
The HelmOps reconciler would previously only update a HelmOp's
`Accepted` condition in case of an error, meaning that that condition
would stay false even after an update, for instance setting the chart
version to a valid constraint resolving to one or more existing
versions.

Such updates now reset the `Accepted` condition on a HelmOps, which
reflects its actual state.