Enable HelmOps OCI chart rendering
When installing a Helm chart through an OCI reference, HelmOps are
expected to store that reference in the `repo` field, unlike GitOps Helm
options which use the `chart` field to that effect.

Since we do not want to break compatibility with existing gitOps setups,
chart URL resolution and contents rendering is now a bit more flexible
than it used to be, supporting both variants.

By the same token, reporting of failures to install charts is improved
in the HelmOps status, which previously skipped status updates when the
`chart` field was empty.