Enforce validation rules for Helm options in HelmOps (#3795)
* Validate HelmOps fields in reconciler

Only a subset of all possible combinations of repo, chart and version
fields make sense. The HelmOps reconciler now ensures that unsupported
combinations lead to error messages being output in the HelmOp's status,
without any bundle being created.

* Enable HelmOps OCI chart rendering

When installing a Helm chart through an OCI reference, HelmOps are
expected to store that reference in the `repo` field, unlike GitOps Helm
options which use the `chart` field to that effect.

Since we do not want to break compatibility with existing gitOps setups,
chart URL resolution and contents rendering is now a bit more flexible
than it used to be, supporting both variants.

By the same token, reporting of failures to install charts is improved
in the HelmOps status, which previously skipped status updates when the
`chart` field was empty.

* Add end-to-end test for HelmOps with tarball

A new end-to-end test case validates that HelmOps supports installing a
Helm chart referenced through a tarball.