Check Helm for static version
When a HelmOp resource references a static chart version, the HelmOps
controller checks Helm for existence of that version.

This prevents errors about the version not existing in the Helm index
from going undetected until a bundle deployment is created and an agent
attempts to deploy it.
Consequently, this also ensures that errors about the version not
existing are produced even if no target cluster is found for the HelmOp.