Use charts branch `dev-v2.12` when testing Fleet in Rancher (#3860)
Charts repository branch dev-v2.11 is outdated, because our workflow
testing Fleet in Rancher installs the latest Rancher pre-release
(currently v2.12.*).
Using that branch led to errors with a webhook pre-release chart version
pinned for v2.12 not being found in the dev-v2.11 charts branch, which
prevented the check on the rancher-webhook deployment from
succeeding, blocking the workflow for hours.