Check Helm for static version (#3858)
* Check Helm for static version

When a HelmOp resource references a static chart version, the HelmOps
controller checks Helm for existence of that version.

This prevents errors about the version not existing in the Helm index
from going undetected until a bundle deployment is created and an agent
attempts to deploy it.
Consequently, this also ensures that errors about the version not
existing are produced even if no target cluster is found for the HelmOp.

* Remove extra wrapping from error message

When a Helm chart version cannot be found, the error is already wrapped
with a message stating `could not get a chart version`. No need to wrap
it with a similar message.

* Reduce risk of conflicts in HelmOps status updates

The HelmOps status reconciler now uses a `Patch` operation, updating
only changed status fields as opposed to replacing a full HelmOps status
subresource.

* Fix `GotNewChartVersion` visibility issue

The condition on emitting the event was wrong, as a new, not yet found
version should be determined based on the status of the HelmOp, which is
later updated by the same polling job.