Use the upstream connection to read helm secrets in the agent
When installing HelmOps in downstream agents (different from the local one) we had an issue because the agent was trying to read the secret from the downstream client.

The secret is cloned to the cluster's namespace in the upstream cluster so, same as we do when reading the BundleDeployment, we should read the secret from upstream.

Also the functions affected now use a `client.Reader` which is safer than `client.Client` because Readers only implement the `Get` and `List` functions.

Refers to: https://github.com/rancher/fleet/issues/3658

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>