Test race condition resulting in missing resources
When a GitRepo is updated such that a path to a bundle changes, but
resources overlap between the old and new paths, there is a risk of
obsolete bundles being cleaned up after newer bundles have been created.
This may happen intermittently due to a race condition, and cause
deletion of resources from the newer bundle, leading to the GitRepo
status appearing as _Modified_.

This commit adds end-to-end test cases exposing the issue, to make it
easier to troubleshoot.