Delete obsolete bundles before writing new ones
When a GitRepo is updated such that a path to a bundle changes, but
resources overlap between the old and new paths, there was a risk of
obsolete bundles being cleaned up after newer bundles had been created.
This would happen intermittently due to a race condition, and cause
deletion of resources from the newer bundle, leading to the GitRepo
status appearing as _Modified_.

To remedy this, `fleet apply` now deletes obsolete bundles synchronously
before writing new ones, that last step happening asynchronously as it
previously did.