Wait for cattle-fleet-local-system namespace existence
When running the CI workflow upgrading Fleet standalone, the
`cattle-fleet-local-system` namespace may not have been created by the
time `deploy-fleet.sh` checks the rollout status of the Fleet agent in
that namespace.
The deployment script now waits until that namespace exists, to prevent
failures in case it does not on the first attempt.