Add support for GitHub App authentication (#3809)
This enables Fleet to authenticate as a Github App installation to
interact with git repositories, besides already supported SSH and basic
auth mechanisms.
Using GitHub apps has many advantages, including not being tied to a
user, higher rate limits and easier/tighter scoping.

To enable this, users must create a secret in the GitRepo's namespace,
as they would for SSH or basic auth. The secret must contain the
following fields:
* `github_app_id`, numerical value
* `github_app_installation_id`, numerical value
* `github_app_private_key`

If the type of the client secret referenced in a GitRepo is neither
basic authentication nor SSH, Fleet will look for the above mentioned
expected keys inside the secret. Based on that, it will use
`ghinstallation` [1] to retrieve a token from GitHub that will be used
for subsequent operations.

[1]: https://github.com/bradleyfalzon/ghinstallation

---------

Co-authored-by: Corentin Néau <tan.neau@suse.com>