[v0.13] Fix CI blockers (#4191)
* Swap bitnami for SUSE registry for codegen helper image

`bitnami/python:3.10` no longer exists on DockerHub.

* Fix typos

Those surfaced when we updated some Github actions, and are unrelated to
recent changes.

* Wait for cattle-fleet-local-system namespace existence

When running the CI workflow upgrading Fleet standalone, the
`cattle-fleet-local-system` namespace may not have been created by the
time `deploy-fleet.sh` checks the rollout status of the Fleet agent in
that namespace.
The deployment script now waits until that namespace exists, to prevent
failures in case it does not on the first attempt.