Support git auth with access token (#4190)
Fleet is now able to authenticate to git using an access token, beside
HTTP basic auth and SSH.
This applies to fetching a repository's latest commit and to cloning.
While Github supports personal access tokens on their own, Bitbucket
requires a username to be set as `x-token-auth` beside the token.