Treat options secret CRUD errors as retryable
When the bundle reconciler fails to create, update or delete an options
secret, that error comes from the Kubernetes API server, and therefore
leads to retries, without the user being able to (nor needing to) do
anything to mitigate the error.
Therefore, Fleet simply retries in such cases, and logs the error
instead of propagating it to the bundle status.