Avoids returning an error when deleting OCI artifacts (#4221)
* Avoids returning an error when deleting OCI artifacts

Returning an error when deleting an OCI artifact could lead to infinite requeing if there is a permanent
error accessing the registry.

The registry is an external resource out of Fleet and we don't have full control.
We already log an event when there is an error deleting the OCI artifact.

This was leading to errors in e2e tests, because one of the tests was checking if there was an error
when trying to use credentials with no delete permissions for the OCI registry.

If the test was the last one, it was logging the expected event and CI was successful, otherwise the next
test was failing randomly depending on which test was (if the test was reseting the OCI storage env variable then
the controller was restarted and unblocked the situation).


Orders the tests and now the only test that needs the OCI_STORAGE=false env variable is executed the last one.




---------

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>