Copy downstream resources from cluster to downstream target namespace
The agent is now able to read a bundle deployment's
`DownstreamResources` field and to copy the corresponding resources into
the target namespace for the Helm release created when installing the
bundle deployment.

Upon deletion of the bundle deployment, the agent will delete those
resources, unless `keepResources` is set to `true` on the bundle
deployment.