Reduce flakiness of signal handling tests
End-to-end tests for signal handling depend on availability of
deployments, which may be affected by other running test suites, as is
the case of the Fleet agent with pod disruption budgets and priority
classes.
Therefore, signal handling tests now retry killing a deployment if that
deployment or its targeted container is not available.