Re-enable linters (#4263)
* Revert "No need to manually enable linters"

This reverts commit 25596d7119ff99abe1e7cea9cdefe45077864583.

* Ignore duplicates raised by linter

Marked blocks are not actually duplicate, but rather common patterns:
* metrics are defined using similar-looking logic across resource types
* the agent uses similar, but not identical, logic to copy config maps
  and secrets from the upstream cluster. Attempts have been made at
  deduplicating that logic with generics, but the need to mutate
  different fields for each resource type made that harder.

* Remove parameter linter errors in unit tests

This removes warnings by:
* adding a `nolint:unparam` line where the parameter which has the same
  value throughout all calls is a string (port or namespace), as that
  value may change in future test cases, or is better specified at a
  single location to make tests easier to maintain
* removing that parameter in one instance, where it was a boolean which
  could easily be added back as a parameter if needed.

* Fix size assertion in integration tests

Expecting a slice to be empty instead of having a length of 0 makes the
linter happy.

* Remove unused parameter from helmOp validation

Validating a HelmOp's fields is a local-only operation and therefore
does not need a context.

* Remove unused logger parameter

Deleting a polling job for a helmOp does not log.
Future logging may happen in the caller, when distinguishing returned
errors between retryable and non-retryable ones.

* Ignore linter complexity warnings for existing methods

Both cluster import and bundle reconciles are known to be complex
processes. The latter could be refactored into separate functions, but
not for the mere sake of making the linter happy.

* Ignore int32 warning for priority class value

Fleet's priority class API could be updated to resolve this issue in a
future release.