Use a scheduler for gitops polling (#4257)
* Use a scheduler for gitops polling

This PR replaces the mechanism of using *RequeueAfter* to implement polling with a Quartz Scheduler similar to the one used in *HelmOps*.

An attempt has been made to use the same approach so that the schedulers are as similar as possible; perhaps in a future iteration the scheduler can be further generalized.

A new Quartz trigger (`ControllerTrigger`) has been introduced, which is used in both *GitOps* and *HelmOps*. Besides having the functionality to execute the task when we add the job to the scheduler, it also has the ability to add jitter to avoid collisions in scenarios with many resources.

Some new unit tests have been created, and others have been removed as, given the new implementation, they no longer made sense.

Additionally, a new value has been added to the Status of a GitRepo (`PollingCommit`) so that when the job performing the polling finds a new commit, it can patch only this field, avoiding collisions with other possible Status updates.

Refers to: https://github.com/rancher/fleet/issues/3394

---------

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>