Optimise owner reference setting on options secret
Setting an owner reference pointing to a bundle deployment on its
options secret can be done when creating or updating the secret, if the
bundle deployment already exists. This works even if the bundle
deployment is outdated, to be subsequently updated, because its UID is
immutable.
This could save the bundle controller a few separate secret update
requests when repeatedly updating multiple bundle deployments.