Fix race when creating secrets and Kubernetes jobs
When creating secrets and a Kubernetes Job that uses those secrets in
volumes, it is possible that the job is created before the secret has
been synchronized. If the job required the content of the secret to be
mounted, the file would be missing.

Also fixes usage of `controllerutil.CreateOrUpdate`.