Create options secret before bundle deployment (#4284)
* Create options secret before bundle deployment

Based on an idea from #3889 by Alejandro Ruiz, the bundle controller now
creates an options secret, if such a secret is needed, before the bundle
deployment which will end up owning it.
This should prevent some errors about missing secrets.

* Optimise owner reference setting on options secret

Setting an owner reference pointing to a bundle deployment on its
options secret can be done when creating or updating the secret, if the
bundle deployment already exists. This works even if the bundle
deployment is outdated, to be subsequently updated, because its UID is
immutable.
This could save the bundle controller a few separate secret update
requests when repeatedly updating multiple bundle deployments.
---------

Co-authored-by: Alejandro Ruiz <alejandro.ruiz@suse.com>