Make go-getter use CABundles
Whether defined in secrets from Rancher or in a GitRepo resource,
go-getter (`helm.chart` field of `fleet.yaml`) is supposed to use the
certificates specified. First the certificate specified in the GitRepo
resource, then the ones from Rancher. It also should honor
`GitRepo.Spec.insecureSkipTLSVerify` (even when using the Rancher
certificates).

Refers to #3646