Detect overlaps between deleted and created bundles (#4157)
* Test race condition resulting in missing resources

When a GitRepo is updated such that a path to a bundle changes, but
resources overlap between the old and new paths, there is a risk of
obsolete bundles being cleaned up after newer bundles have been created.
This may happen intermittently due to a race condition, and cause
deletion of resources from the newer bundle, leading to the GitRepo
status appearing as _Modified_.

This commit adds end-to-end test cases exposing the issue, to make it
easier to troubleshoot.

* Detect overlaps between bundles of a same GitRepo

`fleet apply` is now able to detect resourceoverlaps, by kind and name,
between:
* obsolete bundles, which will be deleted because they no longer match
  any path in the GitRepo
* new bundles coming from the GitRepo, to be created

When such overlaps are detected, the bundle to be written has a newly
populated `Overwrites` field, which is propagated to bundle deployments
when targeting that bundle, specifying which resources (name, kind and
namespace) this bundle deployment will overwrite from other bundles.

This enables the agent to detect if a bundle deployment's
`ModifiedStatus` features missing resources which may have been deleted
when deleting a previous, obsolete bundle, for the bundle deployment to
be redeployed.

As a side note, it is unclear whether namespaces can even be reliably
populated in bundle deployment overwrites, as a bundle resource's actual
target namespace may depend on multiple factors.

* Delete Helm release before triggering new deployment

In order to enable the agent to redeploy a bundle deployment with
missing, overlapped resources, the corresponding Helm release must be
deleted first, so that a new reconcile loop runs the deploy operation.