Fix race when creating secrets and Kubernetes jobs (#4319)
* Fix race when creating secrets and Kubernetes jobs

When creating secrets and a Kubernetes Job that uses those secrets in
volumes, it is possible that the job is created before the secret has
been synchronized. If the job required the content of the secret to be
mounted, the file would be missing.

Also fixes usage of `controllerutil.CreateOrUpdate`.

* Apply suggestion from @weyfonk

Co-authored-by: Corentin Néau <tan.neau@suse.com>

---------

Co-authored-by: Corentin Néau <tan.neau@suse.com>
(cherry picked from commit 98965d7fd339154d8e5c60286ea97fbcea720d2b)