Adds support for Helm lookup (#4302) (#4380)
* Adds support for Helm lookup

Adds support for using templates with the _lookup_ command from _Helm_ as it was previously not working
because of the first run that we do for the _install_ command we do it in _dryRun_ mode to check for errors.

In that case the Helm command internally was not doing the call to the cluster and the _lookup_ calls where resolving to `nil`.

This PR adds a new function that checks if a given _Chart_ has templates with the _lookup_ command and, if so, it enables the
`DryRunOption="server"` and also sets the option `ClientOnly` to _false_ so Helm connects the cluster and the _lookup_ does not
return `nil`

Refers to: https://github.com/rancher/fleet/issues/1851

---------

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>