[v0.14] - Fix race when creating secrets and Kubernetes jobs (#4387)
* Fix race when creating secrets and Kubernetes jobs

When creating secrets and a Kubernetes Job that uses those secrets in
volumes, it is possible that the job is created before the secret has
been synchronized. If the job required the content of the secret to be
mounted, the file would be missing.

Also fixes usage of `controllerutil.CreateOrUpdate`.

* Fix failing e2e test

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>

---------

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>
Co-authored-by: Patrick Seidensal <9606095+p-se@users.noreply.github.com>
Co-authored-by: Corentin Néau <tan.neau@suse.com>