Change GitRepo deletion. (#4361) (#4398)
* Change GitRepo deletion.

This PR introduces changes to how GitRepos are deleted, as well as their dependent resources.

Two main changes are introduced:

* The function that deletes the GitRepo has been modified to check for any pending resources, especially *Bundles* and *ImageScans*, since previously the *GitRepo* was removed before these resources were fully deleted.

* A new controller is introduced for *Content* resources to track the number of *BundleDeployments* that reference the same one.
  This controller aims to eliminate the conflicts that occurred when updating or deleting *BundleDeployments*, as they were being updated or removed concurrently while modifying a shared *Content* resource.

With these changes, the goal is to eliminate both update conflicts (which triggered additional Reconcile calls in the controllers) and potential race conditions when deleting and recreating *GitRepos* that still had dependent resources pending deletion.

---------

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>